factor out the node row component
