+++ src/Nodes/index.ts
@@ -1,6 +1,7 @@
 export { default as NewNode } from './NewNode'
 export { default as NewSubNode } from './NewSubNode'
 export { default as ViewNode } from './ViewNode'
+export { default as NodeRow } from './NodeRow'
 export { default as NodesLanding } from './NodesLanding'
 
 export type GunId = string
